Consolidated the two test exes into one
