+++ test/powershell/Language/Scripting/NativeExecution/NativeCommandProcessor.Tests.ps1
@@ -28,12 +28,12 @@ Describe 'native commands lifecycle' -tags 'Feature' {
 Describe "Native Command Processor" -tags "Feature" {
 
     BeforeAll {
-        # Find where test/powershell is so we can find the createchildprocess command relative to it
+        # Find where test/powershell is so we can find the testexe command relative to it
         $powershellTestDir = $PSScriptRoot
         while ($powershellTestDir -notmatch 'test[\\/]powershell$') {
             $powershellTestDir = Split-Path $powershellTestDir
         }
-        $createchildprocess = Join-Path (Split-Path $powershellTestDir) tools/CreateChildProcess/bin/createchildprocess
+        $testexe = Join-Path (Split-Path $powershellTestDir) tools/TestExe/bin/testexe
     }
 
     # If powershell receives a StopProcessing, it should kill the native process and all child processes
@@ -42,20 +42,20 @@ Describe "Native Command Processor" -tags "Feature" {
     It "Should kill native process tree" -Pending {
 
         # make sure no test processes are running
-        # on Linux, the Process class truncates the name so filter using Where-Object
-        Get-Process | Where-Object {$_.Name -like 'createchildproc*'} | Stop-Process
+        Get-Process testexe -ErrorAction SilentlyContinue | Stop-Process
 
         [int] $numToCreate = 2
 
-        $ps = [PowerShell]::Create().AddCommand($createchildprocess)
-        $ps.AddParameter($numToCreate)
+        $ps = [PowerShell]::Create().AddCommand($testexe)
+        $ps.AddArgument("-createchildprocess")
+        $ps.AddArgument($numToCreate)
         $async = $ps.BeginInvoke()
         $ps.InvocationStateInfo.State | Should Be "Running"
 
         [bool] $childrenCreated = $false
         while (-not $childrenCreated)
         {
-            $childprocesses = Get-Process | Where-Object {$_.Name -like 'createchildproc*'}
+            $childprocesses = Get-Process testexe -ErrorAction SilentlyContinue
             if ($childprocesses.count -eq $numToCreate+1)
             {
                 $childrenCreated = $true
@@ -72,7 +72,7 @@ Describe "Native Command Processor" -tags "Feature" {
                 break
             }
         }
-        $childprocesses = Get-Process | Where-Object {$_.Name -like 'createchildproc*'}
+        $childprocesses = Get-Process testexe
         $count = $childprocesses.count
         $childprocesses | Stop-Process
         $count | Should Be 0
