--- test/tools/EchoArgs/EchoArgs.csproj
@@ -1,11 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <Description>Very simple little console class that you can use to see how PowerShell is passing parameters to legacy console apps.</Description>
-    <TargetFramework>netcoreapp2.0</TargetFramework>
-    <AssemblyName>echoargs</AssemblyName>
-    <OutputType>Exe</OutputType>
-    <RuntimeIdentifiers>ubuntu.16.04-x64;ubuntu.14.04-x64;debian.8-x64;centos.7-x64;fedora.24-x64;win7-x86;win7-x64;win81-x64;win10-x64;osx.10.11-x64;osx.10.12-x64;opensuse.13.2-x64;opensuse.42.1-x64</RuntimeIdentifiers>
-  </PropertyGroup>
-
-</Project>
