+++ test/tools/TestExe/TestExe.cs
@@ -1,0 +1,47 @@
+﻿using System;
+using System.Threading;
+using System.Diagnostics;
+
+namespace TestExe
+{
+    class TestExe
+    {
+        static void Main(string[] args)
+        {
+            if (args.Length > 0)
+            {
+                switch(args[0])
+                {
+                    case "-echoargs":
+                        for (int i = 1; i < args.Length; i++)
+                        {
+                            Console.WriteLine("Arg {0} is <{1}>", i-1, args[i]);
+                        }
+                        break;
+                    case "-createchildprocess":
+                        if (args.Length > 1)
+                        {
+                            uint num = UInt32.Parse(args[1]);
+                            for (uint i = 0; i < num; i++)
+                            {
+                                Process child = new Process();
+                                child.StartInfo.FileName = Process.GetCurrentProcess().MainModule.FileName;
+                                child.StartInfo.Arguments = "-createchildprocess";
+                                child.Start();
+                            }
+                        }
+                        // sleep is needed so the process doesn't exit before the test case kill it
+                        Thread.Sleep(100000);
+                        break;
+                    default:
+                        Console.WriteLine("Unknown test {0}", args[0]);
+                        break;
+                }
+            }
+            else
+            {
+                Console.WriteLine("Test not specified");
+            }
+        }
+    }
+}
